travis: build with 'gradlew build' & try fixing github token use
the default build command is './gradlew check', but i think i want
'./gradlew build' to also run the 'assemble' step, which i believe
makes the jar.

also, i'm passing the github oauth token as an encrypted variable,
which i think means it has to be in a 'secure:' subitem.  i thought
i'd done that before.