retries: extract B2RetryPolicy so it can be replaced.
B2DefaultRetryPolicy implements the default policy which should be fine
for almost all sdk users.  So far, B2StorageClientImpl always uses the
default policy.  I'll work on that in an upcoming change.

Prior to this change, the B2Sleeper was allowed to throw after the
decision was made to sleep some and retry.  As part of extracting the
retry policy, I wanted to be sure that the retry policy would always
find out if the operation succeeded or failed and that wouldn't be the
case if the sleep threw.  I no longer allow the B2Sleeper to throw.
Instead, if it's interrupted, it just resets the thread's interrupted
flag and returns early.  We count on the interruption being noticed
elsewhere (such as in the network call to the B2 servers).

tested with "gradle build javadoc".