retryPolicy: break out tests for B2DefaultRetryPolicy.
previously, B2RetryerTest combined both the tests for the retryer and
for the default retryer policy because they used to be all in the
retryer.  those tests were fine for ensuring that i didn't change the
behavior.

now, i'm testing the two things separately and with some extra
coverage for the new calls to the policy.

tested with unit tests.  (with clean local build & javadoc)