uploading: detect high-level failure to upload a part.
even if there's no transport-level exception, there might be a
semantic-level or unexpected, issue.  now we catch those too.

exercised by added some code to the uploader to throw an exception
after the upload succeeded (as if it had gotten an error back from
the server) and watching B2Sample run.