httpClientFactory: explicitly set the user-agent on the HttpClient
we compute default the HttpClient user-agent the same way HttpClient
does it and we keep the result as a constant, so that we can prevent
HttpClientBuilder doing that work every time we make an HttpClient.
(we make them and discard them constantly and HttpClient computing
the default is expensive because it opens a resource from its jar
and we'd like to avoid those unnecessary syscalls and unzipping.)

besides, we override the value later.  :)