b2Json: fix spelling of "validDurationInSeconds"
it was missing the "In".

i also fixed a similarish misspelling in a local variable for get_download_authorization.
i also added a call to setup the clock as a simulator in a test.  without that change the
clock was being created as a real clock, not a simulator by that test and it was messing
up other tests that expected it to be a clock simulator.  (there could be more of these,
but this is the one i found.)

built & tested with "./gradlew build javadoc writeNewPom".

i also ran the sample.  as expected, it failed because the servers are
currently expecting the misspelling.  i commented that section of the sample
out and everything else succeeded.