Add B2JsonUnionTypeMap.
This replaces the generic Map<> used to hold the mapping from
type name to class.  It's better for type safety, because using
reflection to call methods doesn't let you check the generic type
parameters.

Includes a builder type to make it easy to make one in a static
method.

Testing:
- All Tests