tests/clocks: make sure the unit tests all use a simulated clock.
i was getting a failure recently when running tests from the command
line.  there was an unfortunate ordering which meant sometimes we
ended up with a real clock instead of a simulator and then when a test
that wanted a clock simulator found a real clock, it rightfully failed.

the new B2BaseTest has an @BeforeClass method that ensures that
we'll always have a clock simulator.

tested locally by running "./gradlew build javadoc writeNewPom"