uploads: add optional B2ContentSource.createContentSourceWithRangeOrNull()
This change teaches B2LargeFileUploader to ask B2ContentSource for
a B2ContentSource for a range of its content.  If the source returns
null (the default), we use the same B2PartOfContentSource to
"manually" excerpt, as we've always done.  The new way makes it more
reasonable to use a remote URL as the content source by making it possible
to do a range request on that URL instead of reading and discarding the
start of the URL multiple times.

testing:
  * some additions to some unit tests.
    i'm open to proposals for additional tests.
  * exercised with the new UploadLargeFileFromUrl sample.