add B2StorageClientFactory: a way to create a B2StorageClient without directly naming an implementation.
we're working on adding an OkHttp-based implementation of B2WebApiClient
and we don't want to have the sample programs have to import a specific
web client implementation.  instead, they can now just say:

   B2StorageClientFactory.createDefaultFactory().create(USER_AGENT)

to create a B2StorageClient with whichever of the known implementations
are in the classpath.  in the future, we can allow people to register
their own classes, if that makes sense.

testing:
  * ./gradlew build javadoc writeNewPom
  * ran the unit tests in IDEA, including with coverage to check new tests.
  * ran B2 and B2Sample to make sure clients were created and usable.
  * Mark will uncomment (and maybe tweak) the OkHttp class registration
    in B2StorageClientFactoryPathBasedImpl then make sure that the sample
    programs work with his new implementation without having to change
    the sample code.