Fix bug with Sets contained in unions.
When the `B2JsonUnionBaseHandler` gets the handler for a field, it
needs to use the same logic that `B2JsonObjectHandler` uses to create
the handler for a field.

This worked before the change to allow recursive unions because the fields
were pulled from the object handlers, which had already been initialized
when the union was set up.  The union can no longer assume that, so it
needs to get the un-initialized the field handlers itself.

Testing: new unit test