Fix bug that can leave an uninitialized handler in the map.
If the initialize() method for a handler threw an exception, the
handler was not initialized, but the getHandler() method left it
in the map.  There's now a test for that case, and some cleanup
code to remove handlers from the map if there's an error.

Also, add a test for HashSet being disallowed.

Testing: unit tests