Extend B2FileVersion and B2Part to include contentMd5.
B2 will soon be returning these fields, and it will be good to let
client code see them.  The first code that needs them will be internal
Backblaze tests that exercise the B2 APIs.

B2Headers now has a getContentMd5() method, which reads the "Content-MD5"
header that will be returned when downloading a file.  This header is
also used when getting file info by name; the info is returned in the
headers from a HEAD request on the file.

Testing: All Tests