Disable URI normalization in Apache HttpClient (#106)
By default, HttpClient will normalize URIs, which includes replacing double slashes (`//`) with a single slash. This will alter the request URI and implicitly change the name of files requested through the DownloadByName API where the filename is part of the URI. We need to disable this normalization to allow filename requests to go through as-is.

Updating the HttpClient version to 4.5.9 which is the first version that the setNormalizeUri option became available.