Remove the deprecated okhttp client
This update removes the deprecated okhttp client.  The Apache Commons HttpClient is still included as an implementation fo the B2WebApiClient.

Additionally, a fix for the gradle build in the samples directory is included.  Also fixed the check_code python script to use the core/src/main directory as its default directory to check.

Tested: built & ran unit tests