update version attribute in gradle.properties for b2-sdk-java-private (#15)
This PR updates the version attribute in gradle.properties for b2-sdk-java-private. It currently has 5.1.0-PRIVATE+01. However the Github action for the build adds the build number to the end, so the version in maven has a number such as 5.1.0-PRIVATE+01+76. So this PR removes the +01.

Testing: will view the build number after this PR is built/merged

Jira Ticket: DEV-12614