reduce lock contention in B2Json (#36)
This PR reduces lock contention in B2Json, which is the largest source of lock contention I've seen on api servers. The contention also exists to a lesser degree on pods. Please see Jira Ticket for details.

The main fix is in B2JsonHandlerMap.getHandlerMap to not synchronize the whole method, but only synchronize if the handler is not already in a new map that has only handlers that are fully initialized.

Testing: published new version of b2 java sdk private to local maven and tested bzWebTest basic in the bzmono branch. Also, prior to any changes, I made a small test case in S3ObjectTest and B2FileLockTest to convert multiple dynamic configs to and from B2Json in a loop, then ran them concurrently while I profiled. I then did the same test again after the update and saw the contention was reduced.

Jira Ticket: DEV-17289