DEV-22401 Update B2Json fromJson methods to utilize BufferedReader (#51)
Update B2Json fromJson methods to utilizing BufferedReader to increase performance and decrease objects being created. Previously when only using a InputStreamReader, it would allocate a lot of new char[2] when processing the input stream. Using a BufferedReader, it will allocate a 8k char array (and more as needed), which increases performance and decreases objects allocated.

Tested with unit tests