B2Json.optional[WithDefault] now support omitZero=true
this helps minimize the size of serialized objects that contain a lot of fields whose value is zero.  omitNull=true isn't a great choice for primitive, numeric types because using it means we have to store unboxed values on the heap (taking more memory) and because it complicates use of the fields.

testing:
  * build which runs all the unit tests, including the new test cases.